Crate wit_bindgen

Bindings generation support for Rust with the Component Model.

This crate is a bindings generator for WIT and the Component Model. Users are likely interested in the generate! macro which actually generates bindings. Otherwise this crate provides any runtime support necessary for the macro-generated code.

Macros

generate

Generate bindings for an input WIT document.